fixed waitfg

+++ tsh.c
@@ -200,33 +200,32 @@ void eval(char *cmdline)
 		return;					/* Ignore empty lines */
 	}
 
-    
-	
 	if (!builtin_cmd(argv)) {
         sigemptyset(&mask);
         sigaddset(&mask, SIGCHLD);
         sigprocmask(SIG_BLOCK, &mask, &prev_one); /* Block SIGCHLD */
 		if ((pid = Fork()) == 0) { /* Child runs user job */ 
+            //get new group for child process
             setpgid(0, 0);
             sigprocmask(SIG_SETMASK, &prev_one, NULL); /* Unblock SIGCHLD */
-
 			if (execve(argv[0], argv, environ) < 0) {
 			    printf("%s: Command not found.\n", argv[0]);
 			    exit(0);	
 			}
-            
 		} else { /* Parent */
-            
             /* Parent waits for foreground job to terminate */
-            if (!bg) {
-                addjob(jobs, pid, FG, cmdline);
+            
+            //add child to joblist
+            addjob(jobs, pid,(2 - !bg), cmdline);
+            sigprocmask(SIG_SETMASK, &prev_one, NULL); /* Unblock Parent */
+            if (!bg){
+                //wait for fg
                 waitfg(pid);
-            }
+            } 
             else {
-                addjob(jobs, pid, BG, cmdline);
+                //print for bg
                 printf("[%d] (%d) %s", pid2jid(pid), pid, cmdline);
             }
-            sigprocmask(SIG_SETMASK, &prev_one, NULL); /* Unblock Parent */
         }
 	}
 	return;
@@ -321,41 +320,8 @@ int builtin_cmd(char **argv)
  */
 void do_bgfg(char **argv)
 {
-    // struct job_t *job;
-    // int jid;
-    // pid_t pid;
- 
-    // /* Error handling */
-    // if(argv[1] == NULL) {
-    //     printf("%s command requires pid or jid as seccond argument\n");
-    //     return;
-    // }
- 
-    // /* bg == 0 if arg1 == fg, bg == 1 if arg1 == bg */
-    // int bg = 0;
-    // if (!strcmp(argv[0], "bg")) {
-    //     bg = 1;
-    // }
- 
-    // /*
-    //  * the following three if statements read the first
-    //  * character from the seccond element of argv to see
-    //  * whether the seccond element is a
-    //  * jid : argv[1][0] == '%'
-    //  * pid : 0 > argv[1][0] <= 9
-    //  * neither
-    //  */
-    // if (argv[1][0] == '%') { /* jid */
-    //     jid = atoi()
-    // }
-    // else if (argv[1][0] > '0' && argv[1][0] <= '9') { /* pid */
- 
-    // }
-    // else { /* neither */
- 
      
- 
-    return;
+
 }
  
  
@@ -365,17 +331,24 @@ void do_bgfg(char **argv)
  */
 void waitfg(pid_t pid)
 {
-    int status;
-    if(waitpid(pid, &status, 0) < 0) {
-        unix_error("waitfg: waitpid error");
-    }
-    return;
-    // struct job_t *job = getjobpid(jobs, pid);
- 
-    // while (job->state == FG) {
-    //     sleep(1);
+    // int status;
+    // if(waitpid(pid, &status, 0) < 0) {
+    //     unix_error("waitfg: waitpid error");
     // }
     // return;
+   struct job_t* job;
+    job = getjobpid(jobs,pid);
+    //check if pid is valid
+    if(pid == 0){
+        return;
+    }
+    if(job != NULL){
+        //sleep
+        while(pid==fgpid(jobs)){
+            sleep(1);
+        }
+    }
+    return;
 }
 
 /*****************
@@ -394,7 +367,6 @@ void sigchld_handler(int sig)
     pid_t pid;
     int childStatus;
     while ((pid = waitpid(-1, &childStatus, WNOHANG|WUNTRACED)) > 0) {
-        printf("asdfasdfasdasdf");
         fflush(stdout);
         if (WIFEXITED(childStatus)){
             deletejob(jobs, pid);
@@ -402,10 +374,10 @@ void sigchld_handler(int sig)
         else if(WIFSTOPPED(childStatus)) {
             /*Set job state to stopped*/
             getjobpid(jobs, pid)->state = ST;
-            printf("Job [%d] (%d) stopped by signal %d", pid2jid(pid), pid, WSTOPSIG(childStatus));
+            printf("Job [%d] (%d) stopped by signal %d\n", pid2jid(pid), pid, WSTOPSIG(childStatus));
         }
         else if(WIFSIGNALED(childStatus)) {
-            printf("Job [%d] (%d) terminated by signal %d", pid2jid(pid), pid, WTERMSIG(childStatus));
+            printf("Job [%d] (%d) terminated by signal %d\n", pid2jid(pid), pid, WTERMSIG(childStatus));
             deletejob(jobs, pid);
         }
         else {
@@ -425,7 +397,7 @@ void sigint_handler(int sig)
 {
     //get the pid of the process in the forground
     pid_t pid = fgpid(jobs);
-    if (pid > 0 && pid2jid(pid) > 0)
+    if ((pid > 0) && (pid2jid(pid) > 0))
     {
         /*Send a kill SIGINT*/
         kill(-pid, sig);
@@ -442,7 +414,7 @@ void sigtstp_handler(int sig)
 {
     //get the pid of the process in the forground
     pid_t pid = fgpid(jobs);
-    if (pid > 0)
+    if ((pid > 0) && (pid2jid(pid) > 0))
     {
         /*Send a kill SIGTSTP*/
         kill(-pid, sig);
